+++ tests.webpack.js
@@ -1,5 +1,5 @@
 require('es6-promise').polyfill()
-require('whatwg-fetch')
+require('isomorphic-fetch')
 
 const context = require.context('./modules', true, /-test\.js$/)
 context.keys().forEach(context)