+++ webpack.config.js
@@ -1,5 +1,3 @@
-const webpack = require('webpack')
-
 module.exports = {
 
   output: {
@@ -11,16 +9,6 @@ module.exports = {
     loaders: [
       { test: /\.js$/, exclude: /node_modules/, loader: 'babel' }
     ]
-  },
-
-  resolve: {
-    packageAlias: 'browser'
-  },
-
-  plugins: [
-    new webpack.DefinePlugin({
-      'typeof window': JSON.stringify('object')
-    })
-  ]
+  }
 
 }